Move the networked test(s) into a functional/ dir.
