+++ tests/functional/test_openidbaseclient.py
@@ -1,6 +1,4 @@
-#!/usr/bin/python2 -tt
 # -*- coding: utf-8 -*-
-
 """ Test the OpenIdBaseClient. """
 
 
@@ -8,9 +6,8 @@ import os
 import shutil
 import tempfile
 import unittest
-import functools
 
-from nose.exc import SkipTest
+from .functional_test_utils import networked
 
 from fedora.client import FedoraServiceError
 from fedora.client.openidbaseclient import OpenIdBaseClient
@@ -18,29 +15,6 @@ from fedora.client.openidbaseclient import OpenIdBaseClient
 BASE_URL = 'http://127.0.0.1:5000'
 BASE_URL = 'http://209.132.184.188'
 LOGIN_URL = '{}/login/'.format(BASE_URL)
-
-try:
-    from nose.tools.nontrivial import make_decorator
-except ImportError:
-    # It lives here in older versions of nose (el6)
-    from nose.tools import make_decorator
-
-
-def dev_only(func):
-    """A decorator that skips tests if they are not run by a developer.
-
-    This decorator allows us to have tests that a developer might want to
-    run but shouldn't be run by random people building the package.  It
-    also makes non-networked build systems happy.
-
-    """
-    @functools.wraps(func)
-    def newfunc(self, *args, **kw):
-        if not os.environ.get('DEVELOPER_TESTS', None):
-            raise SkipTest('Only run if you are a python-fedora developer')
-        return func(self, *args, **kw)
-    return make_decorator(func)(newfunc)
-
 
 class OpenIdBaseClientTest(unittest.TestCase):
 
@@ -77,12 +51,12 @@ class OpenIdBaseClientTest(unittest.TestCase):
         if self.saved_session_file:
             shutil.move(self.saved_session_file, self.session_file)
 
-    @dev_only
+    @networked
     def test_no_openid_session(self):
         """Raise FedoraServiceError for no session on service or openid server."""
         self.assertRaises(FedoraServiceError, self.client.login, 'username', 'password')
 
-    @dev_only
+    @networked
     def test_no_service_session(self):
         """Open a service session when we have an openid session."""
         pass
